rename utility/file.ts -> store/utility.ts
